refactor(mysql): scope test support to dedicated modules
