+++ src/infra/adapters/mysql/cli/mod.rs
@@ -15,19 +15,16 @@ pub(super) use policy::{
     validate_mysql_export_query, validate_mysql_multi_query,
     validate_mysql_statements_for_execution,
 };
+#[cfg(feature = "test-support")]
+pub(super) use probe::run_mysql_command_with_timeout;
 pub(super) use probe::{check_mysql_cli_version, probe_mysql_server};
+#[cfg(feature = "test-support")]
+pub(super) use process::run_mysql_adhoc_with_program_and_statements_and_expected_columns;
 pub(super) use process::{
     MYSQL_QUERY_TIMEOUT, MySqlMetadataSession, run_mysql_adhoc, run_mysql_single_statement,
 };
 
-#[cfg(feature = "test-support")]
-pub(super) use process::run_mysql_adhoc_with_timeout_for_test;
 pub(super) use xml::MySqlResultSet;
 
-#[cfg(all(unix, feature = "test-support"))]
-pub(super) use process::run_mysql_cli_script_for_test;
-
-#[cfg(test)]
-pub(super) mod test_support {
-    pub(in crate::adapters::mysql) use super::process::MySqlProcess;
-}
+#[cfg(feature = "test-support")]
+pub(super) mod test_support;
