+++ src/infra/adapters/mysql/executor.rs
@@ -21,105 +21,6 @@ use super::dsn::parse_and_validate_mysql_dsn;
 use super::metadata;
 use super::option_file::MySqlOptionFile;
 
-#[cfg(feature = "test-support")]
-pub(super) mod test_support {
-    use std::path::{Path, PathBuf};
-    use std::time::Duration;
-
-    use crate::adapters::csv_export::export_to_path;
-    use crate::app::ports::outbound::{AccessMode, DbOperationError};
-
-    use super::super::cli::run_mysql_adhoc_with_timeout_for_test;
-    use super::{
-        MySqlOptionFile, export_mysql_csv_to_file, parse_and_validate_mysql_dsn, run_mysql_adhoc,
-        validate_mysql_multi_query,
-    };
-
-    #[doc(hidden)]
-    pub struct MySqlOptionFileForTest {
-        option_file: MySqlOptionFile,
-    }
-
-    impl MySqlOptionFileForTest {
-        #[must_use]
-        pub fn path(&self) -> &Path {
-            &self.option_file.path
-        }
-    }
-
-    #[doc(hidden)]
-    pub fn create_mysql_option_file_for_test(
-        dsn: &str,
-    ) -> Result<MySqlOptionFileForTest, DbOperationError> {
-        let target = parse_and_validate_mysql_dsn(dsn)?;
-        Ok(MySqlOptionFileForTest {
-            option_file: MySqlOptionFile::create(&target)?,
-        })
-    }
-
-    #[doc(hidden)]
-    /// Runs the normal adhoc CLI process with a read-only session without the app-side policy
-    /// gate so integration tests can verify server-side rejection of a side effect.
-    pub async fn execute_mysql_adhoc_with_read_only_session_for_test(
-        dsn: &str,
-        query: &str,
-    ) -> Result<(), DbOperationError> {
-        let target = parse_and_validate_mysql_dsn(dsn)?;
-        let statements =
-            validate_mysql_multi_query(query, target.database.as_deref(), AccessMode::ReadWrite)?;
-        let option_file = MySqlOptionFile::create(&target)?;
-        let result = run_mysql_adhoc(&option_file.path, &statements, AccessMode::ReadOnly).await;
-        drop(option_file);
-        result.map(|_| ())
-    }
-
-    #[doc(hidden)]
-    pub async fn execute_mysql_adhoc_with_timeout_for_test(
-        dsn: &str,
-        query: &str,
-        execution_timeout: Duration,
-    ) -> Result<(), DbOperationError> {
-        let target = parse_and_validate_mysql_dsn(dsn)?;
-        let statements =
-            validate_mysql_multi_query(query, target.database.as_deref(), AccessMode::ReadWrite)?;
-        let option_file = MySqlOptionFile::create(&target)?;
-        let result = run_mysql_adhoc_with_timeout_for_test(
-            &option_file.path,
-            &statements,
-            AccessMode::ReadWrite,
-            execution_timeout,
-        )
-        .await;
-        drop(option_file);
-        result.map(|_| ())
-    }
-
-    #[cfg(unix)]
-    #[doc(hidden)]
-    pub async fn run_mysql_cli_script_for_test(
-        dsn: &str,
-        script: &str,
-    ) -> Result<Vec<u8>, DbOperationError> {
-        super::super::cli::run_mysql_cli_script_for_test(dsn, script).await
-    }
-
-    #[doc(hidden)]
-    /// Runs the export process without client-side query policy validation so integration tests can
-    /// verify that the MySQL read-only session rejects a side effect at the server boundary.
-    pub async fn export_mysql_csv_to_path_for_test(
-        dsn: &str,
-        query: &str,
-        path: PathBuf,
-    ) -> Result<PathBuf, DbOperationError> {
-        let target = parse_and_validate_mysql_dsn(dsn)?;
-        let query = query.to_string();
-        export_to_path(path, move |temporary_path| async move {
-            export_mysql_csv_to_file(target, &query, temporary_path).await
-        })
-        .await
-    }
-}
-
 async fn execute_adhoc_with_statements(
     dsn: &str,
     query: &str,
